pyproject.toml
[project]
name = "mm-std"
version = "0.0.1"
description = ""
requires-python = ">=3.11"
dependencies = [
    "requests~=2.32.3",
    "PySocks~=1.7.1",
    "pydash~=8.0.0",
    "python-dateutil~=2.9.0",
    "pydantic~=2.8.2",
    "python-dotenv~=1.0.1",
    "PyYAML~=6.0.1",
    "cryptography~=43.0.0",
    "rich",
]
[project.optional-dependencies]
dev = [
    "build~=1.2.1",
    "twine~=5.1.0",
    "pytest~=8.3.2",
    "pytest-xdist~=3.6.1",
    "pytest-httpserver~=1.0.8",
    "coverage~=7.6.0",
    "ruff~=0.5.2",
    "pip-audit~=2.7.0",
    "bandit~=1.7.7",
    "mypy~=1.11.0",
    "types-python-dateutil~=2.9.0",
    "types-requests~=2.32.0.20240523",
    "types-PyYAML~=6.0.12.12",
]


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[tool.mypy]
python_version = "3.12"
warn_no_return = false
strict = true
exclude = ["^tests/", "^tmp/"]


[tool.ruff]
line-length = 130
target-version = "py312"
lint.select = [
    "F", # Pyflakes
    "E", "W", # pycodestyle
    "B", # flake8-bugbear
    "A", # flake8-builtins
    "COM", # flake8-commas
    "C40", # flake8-comprehensions
    "G", # flake8-logging-format
    "PIE", # flake8-pie
    "T20", # flake8-print
    "RUF", # Ruff-specific rules
]
lint.ignore = [
    "A003", # builtin-attribute-shadowing
    "UP040", # non-pep695-type-alias
    "COM812"
]
[tool.ruff.format]
quote-style = "double"
indent-style = "space"


[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B311"]


[tool.pytest.ini_options]
markers = ["proxy: requires access proxies", "telegram: requires a telegram bot"]
